Change BLAS implementation from accelerate to netlib
Currently, the BLAS implementation defaults to accelerate, the Apple
one, which is macOS only. On top of preventing non macOS user to build
the code as-is, it also causes the CI to fail.

Netlib is the "vanilla" BLAS implementation. Not very efficient, but
should work on the highest number of platforms.